cargo-bench-history-stress: lib+bin split, Miri-clean tests, kill set_index mutants
Adopt the repo's lib.rs+main.rs structure for CLI utilities: all logic
moves into the library crate behind a public `run() -> ExitCode`, and
main.rs becomes a thin coverage-off shim. This also restores coverage CI,
which runs `cargo llvm-cov --lib` and failed on the previously bin-only
crate.

Make the new tests Miri-compatible: the FakeRunner-path async tests perform
no real async I/O, so drive them with `futures::executor::block_on` instead
of a tokio runtime (a tokio reactor cannot initialize under Miri). Tests
that genuinely need the OS (real subprocess, real filesystem, wall-clock)
stay on tokio where needed and are gated `#[cfg_attr(miri, ignore)]`.

Add a set_index test covering every Task variant with values distinct from
0 and 1, killing the two remaining `set_index -> usize` constant-return
mutants.
